Added two button and one Label widget 12/30/2022

+++ BiquOperatorPanel/Pheonix/TouchLCD/Widgets/widget.h
@@ -4,12 +4,12 @@
 
 typedef struct
 {
-	float x, y;
+	uint16_t x, y;
 }Point;
 
 typedef struct
 {
-	float width, height;
+	uint16_t width, height;
 }Size;
 
 typedef enum
